Fix NRE if user immediately completes Deferal
Made the deferral flow more linear and easy to follow
Moved the navigation parts of Shell out to a separate class
