+++ Xamarin.Forms.Core.UnitTests/ShellNavigatingTests.cs
@@ -47,6 +47,69 @@ namespace Xamarin.Forms.Core.UnitTests
 			Assume.That(shell.CurrentState.Location.ToString(), Is.EqualTo("//one/tabone/content"));
 		}
 
+		[Test]
+		public async Task CancelNavigationShellItems()
+		{
+			var flyoutItem = CreateShellItem<FlyoutItem>();
+			TestShell shell = new TestShell()
+			{
+				Items = { flyoutItem }
+			};
+
+			var navigatingToShellContent = CreateShellContent();
+			shell.Items[0].Items[0].Items.Add(navigatingToShellContent);
+
+			bool executed = false;
+			TaskCompletionSource<object> taskCompletionSource = new TaskCompletionSource<object>();
+
+			ShellContent contentActiveBeforeCompletingDeferral = null;
+			shell.Navigating += async (_, args) =>
+			{
+				var deferral = args.GetDeferral();
+				await Task.Delay(100);
+
+				contentActiveBeforeCompletingDeferral = flyoutItem.Items[0].Items[0];
+				executed = true;
+				deferral.Complete();
+			};
+
+			shell.Navigated += (_, args) =>
+			{
+				taskCompletionSource.SetResult(true);
+			};
+
+			shell.Controller.ProposeNavigation(
+				ShellNavigationSource.ShellContentChanged, flyoutItem, flyoutItem.Items[0], navigatingToShellContent, flyoutItem.Items[0].Stack, true);
+
+			await taskCompletionSource.Task;
+
+			Assert.IsTrue(executed);
+			Assert.AreNotEqual(contentActiveBeforeCompletingDeferral, navigatingToShellContent);
+			Assert.AreEqual(flyoutItem.Items[0].Items[0], contentActiveBeforeCompletingDeferral, "Navigation to new Content was not deferred");
+			Assert.AreEqual(flyoutItem.Items[0].CurrentItem, navigatingToShellContent, "Navigation after completing the deferral failed");
+		}
+
+		[Test]
+		public async Task ImmediatelyCompleteDeferral()
+		{
+			TestShell shell = new TestShell()
+			{
+				Items = { CreateShellItem<FlyoutItem>() }
+			};
+
+			bool executed = false;
+			shell.Navigating += (_, args) =>
+			{
+				var deferral = args.GetDeferral();
+				executed = true;
+				deferral.Complete();
+			};
+
+			await shell.Navigation.PushAsync(new ContentPage());
+			Assert.IsTrue(executed);
+			Assert.AreEqual(2, shell.Navigation.NavigationStack.Count);
+		}
+
 		[TestCase("PopToRoot")]
 		[TestCase("Pop")]
 		public async Task DeferPopNavigation(string testCase)
