+++ Xamarin.Forms.Core/Shell/ShellNavigatingEventArgs.cs
@@ -9,9 +9,16 @@ namespace Xamarin.Forms
 		int _deferalCount;
 		Func<Task> _deferralFinishedTask;
 		TaskCompletionSource<bool> _deferredTaskCompletionSource;
+		bool _deferralCompleted = false;
 
 		public ShellNavigatingEventArgs(ShellNavigationState current, ShellNavigationState target, ShellNavigationSource source, bool canCancel)
 		{
+
+#if NETSTANDARD2_0
+			_deferralFinishedTask = () => Task.CompletedTask;
+#else
+			_deferralFinishedTask = () => Task.Delay(0);
+#endif
 			Current = current;
 			Target = target;
 			Source = source;
@@ -40,6 +47,9 @@ namespace Xamarin.Forms
 
 		public ShellNavigatingDeferral GetDeferral()
 		{
+			if (_deferralCompleted)
+				throw new InvalidOperationException("Deferral has already been completed");
+
 			if (!CanCancel)
 				return null;
 
@@ -57,13 +67,14 @@ namespace Xamarin.Forms
 		{
 			if (Interlocked.Decrement(ref _deferalCount) == 0)
 			{
-				var task = _deferralFinishedTask();
-				_deferralFinishedTask = null;
+				_deferralCompleted = true;
 
 				try
 				{
+					var task = _deferralFinishedTask();
+					_deferralFinishedTask = null;
 					await task;
-					_deferredTaskCompletionSource.SetResult(true);
+					_deferredTaskCompletionSource.SetResult(!Cancelled);
 				}
 				catch (TaskCanceledException)
 				{
@@ -83,6 +94,9 @@ namespace Xamarin.Forms
 		internal bool DeferredEventArgs { get; set; }
 
 		internal int DeferralCount => _deferalCount;
+
+		internal bool NavigationDelayedOrCancelled =>
+			Cancelled || DeferralCount > 0;
 
 		internal void RegisterDeferralCompletedCallBack(Func<Task> deferralFinishedTask)
 		{
